chore: updating queries to use `Quickstart`
Previously, the type we were using came from `gatsby-source-filesystem`
and was called `Quickstarts` (notice the extra `s`). This commit updates
all of the queries to use the right type. It also updates them to look
for `slug` as a top-level property rather than in the `fields` property.

This also cleans up the query on the index page. We didn't need to
include all the information about quickstarts, just some basic stuff
like the name and logo.