chore: Query for SVG logos separately
This change, while somewhat clunky in the query, prevents the following
error from showing up during builds and local development:

```
warn You can't use childImageSharp together with logo.svg — use publicURL instead. The childImageSharp portion of the query in this file will return null:
```

The reason for this error is that sharp does not work with SVG files -
they are already performant and do not need any additional
optimizations. We needed a separate field so that we could avoid
querying `childImageSharp`.